requirements.txt
Pillow==11.2.1
numpy==2.1.3
tensorflow-cpu==2.19.0
matplotlib==3.10.3
opencv-python==4.12.0.88